V4

+++ basis/bin/destroy_all.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 if [ "$PROJECT_DIR" == "" ]; then
   echo "ERROR: PROJECT_DIR undefined. Please use starter.sh destroy"
   exit 1
@@ -8,62 +8,17 @@ SECONDS=0
 
 # Call the script with --auto-approve to destroy without prompt
 . starter.sh env -no-auto
-title "OCI Starter - Destroy"
-echo 
-echo "Warning: This will destroy all the resources created by Terraform."
-echo 
-if [ "$1" != "--auto-approve" ]; then
-  read -p "Do you want to proceed? (yes/no) " yn
-
-  case $yn in 
-  	yes ) echo Deleting;;
-	no ) echo Exiting...;
-		exit 1;;
-	* ) echo Invalid response;
-		exit 1;;
-  esac
-fi
-
-. starter.sh env
-
-# Check if there is something to destroy.
-if [ -f $STATE_FILE ]; then
-  export TF_RESOURCE=`cat $STATE_FILE | jq ".resources | length"`
-  if [ "$TF_RESOURCE" == "0" ]; then
-    echo "No resource in terraform state file. Nothing to destroy."
-    exit 0
-  fi
-else
-  echo "File $STATE_FILE does not exist. Nothing to destroy."
-  exit 0
-fi
-
-# before_destroy.sh
-if [ -f src/before_destroy.sh ]; then
-  src/before_destroy.sh
-fi
 
 # Confidential APP
 disableConfidentialApp() {
   # Disable the app before destroy... (Bug?) if not destroy fails...
   CONFIDENTIAL_APP_OCID=$1    
   echo "Confidential app: set active to false. APP_ID=$CONFIDENTIAL_APP_OCID"
-  get_output_from_tfstate "IDCS_URL" "idcs_url"
   # Remove trailing /
   IDCS_URL=${IDCS_URL::-1}
   oci identity-domains app-status-changer put --force --active false --app-status-changer-id $CONFIDENTIAL_APP_OCID --schemas '["urn:ietf:params:scim:schemas:oracle:idcs:AppStatusChanger"]' --endpoint $IDCS_URL  --force
-  exit_on_error
+  exit_on_error "disableConfidentialApp"
 }
-for CONFIDENTIAL_APP_OCID in `cat $STATE_FILE | jq -r '.resources[] | select(.type=="oci_identity_domains_app") | .instances[].attributes.id'`;
-do
-   disableConfidentialApp $CONFIDENTIAL_APP_OCID
-done;
-
-# OKE
-if [ -f $PROJECT_DIR/src/terraform/oke.tf ]; then
-  title "OKE Destroy"
-  $BIN_DIR/oke_destroy.sh --auto-approve
-fi
 
 # Buckets
 cleanBucket() {
@@ -76,22 +31,73 @@ cleanBucket() {
     echo "No Object storage $BUCKET_NAME"
   fi  
 }
-for BUCKET_NAME in `cat $STATE_FILE | jq -r '.resources[] | select(.type=="oci_objectstorage_bucket") | .instances[].attributes.name'`;
-do
-  cleanBucket $BUCKET_NAME
-done;
 
-title "Terraform Destroy"
-$BIN_DIR/terraform_destroy.sh --auto-approve -no-color
-exit_on_error
+if [ "$TF_VAR_infra_as_code" != "from_resource_manager" ]; then
 
-export TF_RESOURCE=`cat $STATE_FILE | jq ".resources | length"`
-if [ "$TF_RESOURCE" == "0" ]; then
-  echo "Empty state file - cleaning up .terraform"
-  rm -Rf $PROJECT_DIR/src/terraform/.terraform
-  rm -Rf $PROJECT_DIR/src/terraform/.terraform.lock.hcl
-  export DESTROY_DATE=`date '+%Y%m%d-%H%M%S'`
-  mv $TARGET_DIR $TARGET_DIR.$DESTROY_DATE
+  # Check if there is something to destroy.
+  title "OCI Starter - Destroy"
+  echo 
+  echo "Warning: This will destroy all the resources created by Terraform."
+  echo 
+  if [ "$1" != "--auto-approve" ] && [ "$1" != "--called_by_resource_manager" ]; then
+    read -p "Do you want to proceed? (yes/no) " yn
+    case $yn in 
+      yes ) echo Deleting;;
+      no ) echo Exiting...;
+           exit 1;;
+      * ) echo Invalid response;
+          exit 1;;
+    esac
+  fi
+  . starter.sh env
+
+  # Check if there is something to destroy.
+  if [ -f $STATE_FILE ]; then
+    export TF_RESOURCE=`cat $STATE_FILE | jq ".resources | length"`
+    if [ "$TF_RESOURCE" == "0" ]; then
+        echo "No resource in terraform state file. Nothing to destroy."
+        exit 0
+    fi
+  else
+    echo "File $STATE_FILE does not exist. Nothing to destroy."
+    exit 0
+  fi
+
+  # before_destroy.sh
+  if [ -f src/before_destroy.sh ]; then
+    src/before_destroy.sh
+  fi
+
+  for CONFIDENTIAL_APP_OCID in `cat $STATE_FILE | jq -r '.resources[] | select(.type=="oci_identity_domains_app") | .instances[].attributes.id'`;
+  do
+    disableConfidentialApp $CONFIDENTIAL_APP_OCID
+  done;
+
+    # OKE
+  if [ -f $PROJECT_DIR/src/terraform/oke.tf ]; then
+    title "OKE Destroy"
+    $BIN_DIR/destroy_oke.sh --auto-approve
+  fi
+
+  for BUCKET_NAME in `cat $STATE_FILE | jq -r '.resources[] | select(.type=="oci_objectstorage_bucket") | .instances[].attributes.name'`;
+  do
+    cleanBucket $BUCKET_NAME
+  done;
 fi
 
-echo "Destroy time: ${SECONDS} secs"
+if [ "$1" != "--called_by_resource_manager" ]; then
+  title "Terraform Destroy"
+  $BIN_DIR/terraform_destroy.sh --auto-approve -no-color
+  exit_on_error "terraform_destroy.sh"
+
+  export TF_RESOURCE=`cat $STATE_FILE | jq ".resources | length"`
+  if [ "$TF_RESOURCE" == "0" ]; then
+    echo "Empty state file - cleaning up .terraform"
+    rm -Rf $PROJECT_DIR/src/terraform/.terraform
+    rm -Rf $PROJECT_DIR/src/terraform/.terraform.lock.hcl
+    export DESTROY_DATE=`date '+%Y%m%d-%H%M%S'`
+    mv $TARGET_DIR $TARGET_DIR.$DESTROY_DATE
+  fi
+
+  echo "Destroy time: ${SECONDS} secs"
+fi
